+++ src/main/java/com/dabomstew/pkrandom/Randomizer.java
@@ -560,39 +560,42 @@ public class Randomizer {
                 }
                 romHandler.setStarters(starters);
                 romHandler.getTemplateData().put("startersList", starters);
-            } else if (settings.getStartersMod() == Settings.StartersMod.COMPLETELY_RANDOM) {
-                // Randomise
-                int starterCount = romHandler.isYellow() ? 2 : 3;
-                List<Pokemon> starters = new ArrayList<Pokemon>();
-                selectRandomStarter(starterCount, starters, () -> romHandler.randomStarterPokemon(
-                    settings.isStartersNoSplit(), settings.isStartersUniqueTypes(), 
-                    settings.isStartersBaseEvoOnly(),
-                    settings.isStartersLimitBST() ? settings.getStartersBSTLimitModifier() : 9999));
+            } else if (settings.getStartersMod() == Settings.StartersMod.RANDOM) {
+                switch (settings.getStartersMinimumEvos()) {
+                    case 0:
+                        int starterCount = romHandler.isYellow() ? 2 : 3;
+                        List<Pokemon> starters = new ArrayList<Pokemon>();
+                        selectRandomStarter(starterCount, starters, () -> romHandler.randomStarterPokemon(
+                            settings.isStartersNoSplit(), settings.isStartersUniqueTypes(), 
+                            settings.isStartersBaseEvoOnly(),
+                            settings.isStartersLimitBST() ? settings.getStartersBSTLimitModifier() : 9999));
 
-                romHandler.setStarters(starters);
-                romHandler.getTemplateData().put("startersList", starters);
-            } else if (settings.getStartersMod() == Settings.StartersMod.RANDOM_WITH_ONE_OR_TWO_EVOLUTIONS) {
-                // Randomise
-                int starterCount = romHandler.isYellow() ? 2 : 3;
-                List<Pokemon> starters = new ArrayList<Pokemon>();
-                selectRandomStarter(starterCount, starters, () -> 
-                    romHandler.random1or2EvosPokemon(settings.isStartersNoSplit(), 
-                    settings.isStartersUniqueTypes(), settings.isStartersBaseEvoOnly(),
-                    settings.isStartersLimitBST() ? settings.getStartersBSTLimitModifier() : 9999));
+                        romHandler.setStarters(starters);
+                        romHandler.getTemplateData().put("startersList", starters);
+                        break;
+                    case 1:            
+                        starterCount = romHandler.isYellow() ? 2 : 3;
+                        starters = new ArrayList<Pokemon>();
+                        selectRandomStarter(starterCount, starters, () -> 
+                            romHandler.random1or2EvosPokemon(settings.isStartersNoSplit(), 
+                            settings.isStartersUniqueTypes(), settings.isStartersBaseEvoOnly(),
+                            settings.isStartersLimitBST() ? settings.getStartersBSTLimitModifier() : 9999));
 
-                romHandler.setStarters(starters);
-                romHandler.getTemplateData().put("startersList", starters);
-            } else if (settings.getStartersMod() == Settings.StartersMod.RANDOM_WITH_TWO_EVOLUTIONS) {
-                // Randomise
-                int starterCount = romHandler.isYellow() ? 2 : 3;
-                List<Pokemon> starters = new ArrayList<Pokemon>();
-                selectRandomStarter(starterCount, starters, () -> 
-                    romHandler.random2EvosPokemon(settings.isStartersNoSplit(), 
-                    settings.isStartersUniqueTypes(), settings.isStartersBaseEvoOnly(),
-                    settings.isStartersLimitBST() ? settings.getStartersBSTLimitModifier() : 9999));
+                        romHandler.setStarters(starters);
+                        romHandler.getTemplateData().put("startersList", starters);
+                        break;
+                    case 2:
+                        starterCount = romHandler.isYellow() ? 2 : 3;
+                        starters = new ArrayList<Pokemon>();
+                        selectRandomStarter(starterCount, starters, () -> 
+                            romHandler.random2EvosPokemon(settings.isStartersNoSplit(), 
+                            settings.isStartersUniqueTypes(), settings.isStartersBaseEvoOnly(),
+                            settings.isStartersLimitBST() ? settings.getStartersBSTLimitModifier() : 9999));
 
-                romHandler.setStarters(starters);
-                romHandler.getTemplateData().put("startersList", starters);
+                        romHandler.setStarters(starters);
+                        romHandler.getTemplateData().put("startersList", starters);
+                        break;
+                }
             }
             if (settings.isRandomizeStartersHeldItems() && !(romHandler instanceof Gen1RomHandler)) {
                 romHandler.randomizeStarterHeldItems(settings.isBanBadRandomStarterHeldItems());